+++ apps/skde/src/components/StyledComponents/PageWrapper.tsx
@@ -5,7 +5,7 @@ export const PageWrapper = ({ children }: PropsWithChildren) => (
   <Box
     sx={{
       backgroundColor: "background.paper",
-      "& .header-top, & .header-middle, & .main-toolbar, & .footer": {
+      "& .header-top, & .header-middle, & .main-toolbar": {
         paddingX: { xs: 2, sm: 4, lg: 6, xl: 16 },
       },
       "& .menu-wrapper": {
@@ -17,7 +17,7 @@ export const PageWrapper = ({ children }: PropsWithChildren) => (
         paddingX: { xs: 2, sm: 4, lg: 6, xl: 16 },
         paddingLeft: { xxl: 2 },
       },
-      "& .hospital-profile-box, & .hospital-profile-footer": {
+      "& .hospital-profile-box": {
         paddingX: { xs: 2, sm: 4, lg: 6, xl: 16, xxl: 0 },
       },
     }}